# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Yantra4D API
  version: 1.0.0
  description: |
    REST API for the Yantra4D parametric 3D design platform.
    Provides OpenSCAD rendering, project management, AI-assisted design,
    Git operations, and analytics.
  license:
    name: AGPL-3.0
    url: https://opensource.org/license/agpl-v3/
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.yantra4d.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Janua-issued JWT token
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
      required: [status, error]
    ProjectSummary:
      type: object
      properties:
        slug:
          type: string
          example: gridfinity
        name:
          type: string
          example: Gridfinity Bin
        description:
          type: object
          properties:
            en:
              type: string
            es:
              type: string
        thumbnail:
          type: string
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        stats:
          type: object
          description: Only present when ?stats=1
          properties:
            renders:
              type: integer
            exports:
              type: integer
            preset_applies:
              type: integer
    Manifest:
      type: object
      properties:
        project:
          type: object
          properties:
            name:
              type: string
            slug:
              type: string
            version:
              type: string
            description:
              type: object
            thumbnail:
              type: string
            tags:
              type: array
              items:
                type: string
            difficulty:
              type: string
        modes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              scad_file:
                type: string
              label:
                type: object
              parts:
                type: array
                items:
                  type: string
              estimate:
                type: object
        parts:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              render_mode:
                type: integer
              label:
                type: object
              default_color:
                type: string
        parameters:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              default:
                oneOf:
                  - type: number
                  - type: string
                  - type: boolean
              min:
                type: number
              max:
                type: number
              step:
                type: number
              label:
                type: object
              visible_in_modes:
                type: array
                items:
                  type: string
              widget:
                type: object
                properties:
                  type:
                    type: string
                    enum: [color-gradient, component-picker]
                  catalog:
                    type: string
                    description: Used when type is component-picker
        estimate_constants:
          type: object
          properties:
            base_time:
              type: number
            per_unit:
              type: number
            per_part:
              type: number
        presets:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              label:
                type: object
              values:
                type: object
      required: [project, modes, parts, parameters, estimate_constants]
    RenderRequest:
      type: object
      properties:
        project:
          type: string
          description: Project slug
        mode:
          type: string
          description: Mode ID from manifest
        scad_file:
          type: string
          description: Deprecated — use mode instead
        parameters:
          type: object
          additionalProperties: true
        export_format:
          type: string
          enum: [stl, 3mf, off]
          default: stl
        ignore_cache:
          type: boolean
          description: Bypass L1 and L2 caches to force a fresh render
          default: false
      required: [mode]
    RenderResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        parts:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
                format: uri
              size_bytes:
                type: integer
        log:
          type: string
...
paths:
...
  /api/projects:
    get:
      tags: [projects]
      summary: List all projects
      operationId: listProjects
      parameters:
        - name: stats
          in: query
          schema:
            type: string
            enum: ["1"]
          description: Include 30-day analytics stats
      responses:
        "200":
          description: Array of project summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectSummary"
  /api/projects/{slug}/manifest:
    get:
      tags: [projects]
      summary: Get project manifest
      operationId: getProjectManifest
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full project manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Manifest"
        "304":
          description: Not modified (ETag match)
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{slug}/meta:
    get:
      tags: [projects]
      summary: Get project metadata
      operationId: getProjectMeta
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project metadata or null
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                  - type: "null"
  /api/projects/{slug}/parts/{filename}:
    get:
      tags: [projects]
      summary: Serve static part file
      operationId: serveStaticPart
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Binary STL file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Path traversal blocked
        "404":
          description: File or project not found
  /api/projects/{slug}/fork:
    post:
      tags: [projects]
      summary: Fork a project
      operationId: forkProject
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new_slug:
                  type: string
                  pattern: "^[a-z0-9][a-z0-9_-]{1,48}[a-z0-9]$"
              required: [new_slug]
      responses:
        "200":
          description: Fork successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  slug:
                    type: string
        "400":
          description: Invalid slug
        "401":
          description: Pro tier required
        "409":
          description: Slug already exists
  /api/projects/{slug}/manifest/assembly-steps:
    put:
      tags: [projects]
      summary: Update assembly steps
      operationId: updateAssemblySteps
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assembly_steps:
                  type: array
                  items:
                    type: object
              required: [assembly_steps]
      responses:
        "200":
          description: Steps saved
        "400":
          description: Missing assembly_steps
        "404":
          description: Project not found

  # ── Legacy Manifest ────────────────────────────────────
...
  /api/render:
    post:
      tags: [render]
      summary: Synchronous render
      operationId: renderStl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderRequest"
      responses:
        "200":
          description: Render complete
          headers:
            X-RateLimit-Limit:
              schema:
                type: string
            X-RateLimit-Tier:
              schema:
                type: string
            X-Cache:
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderResponse"
        "400":
          description: Invalid mode or SCAD file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/projects/{slug}/bom:
    get:
      tags: [bom]
      summary: Get bill of materials
      operationId: getBom
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: BOM data
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
        "404":
          description: Project not found

  # ── Datasheet ───────────────────────────────────────────
  /api/projects/{slug}/datasheet:
    get:
      tags: [datasheet]
      summary: Generate project datasheet
      operationId: generateDatasheet
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [pdf, html]
            default: pdf
        - name: lang
          in: query
          schema:
            type: string
            default: en
      responses:
        "200":
          description: Datasheet document
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
        "404":
          description: Project not found

  # ── Analytics ───────────────────────────────────────────
...
  /api/admin/projects/{slug}:
    get:
      tags: [admin]
      summary: Get project details (admin)
      operationId: adminProjectDetail
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detailed project info
        "404":
          description: Project not found

  # ── Verify ──────────────────────────────────────────────
...